+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Converter framework (`fiboa convert`)
-- Converter for NRW, Germany (`de_nrw`)
+- ...
 
 ### Changed
 
@@ -23,6 +22,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 - ...
+
+### Fixed
+
+- ...
+
+
+## [v0.2.0] - 2024-04-02
+
+### Added
+
+- Converter framework (`fiboa convert`)
+- Converter for NRW, Germany (`de_nrw`)
 
 ### Fixed
 
@@ -78,7 +89,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First release
 
-[Unreleased]: <https://github.com/radiantearth/stac-spec/compare/v0.1.1...main>
+[Unreleased]: <https://github.com/radiantearth/stac-spec/compare/v0.2.0...main>
+[v0.2.0]: <https://github.com/radiantearth/stac-spec/compare/v0.1.1...v0.2.0>
 [v0.1.1]: <https://github.com/radiantearth/stac-spec/compare/v0.1.0...v0.1.1>
 [v0.1.0]: <https://github.com/radiantearth/stac-spec/compare/v0.0.9...v0.1.0>
 [v0.0.9]: <https://github.com/radiantearth/stac-spec/compare/v0.0.8...v0.0.9>